separated HJLogFormat.h
